Simplify toolkit to focus on core benchmarking workflow
Remove transforms/, baselines/, exporters/ modules and utils/graph.py,
utils/io.py. Relocate WindowConfig/SplitConfig to core/types.py,
windowing functions to utils/data.py, and haversine_distance to
utils/geo.py. Add examples/basic_example.py demonstrating the
BenchmarkRunner with LastValueModel.
