+++ core/types.py
@@ -27,7 +27,6 @@ class IRMetadata:
     feature_columns: list[str]
     units: dict[str, str] = field(default_factory=dict)
     source_url: str = ""
-    transform_history: list[str] = field(default_factory=list)
     extra: dict[str, Any] = field(default_factory=dict)
 
     def to_dict(self) -> dict:
@@ -39,7 +38,6 @@ class IRMetadata:
             "feature_columns": self.feature_columns,
             "units": self.units,
             "source_url": self.source_url,
-            "transform_history": self.transform_history,
             "extra": self.extra,
         }
 
@@ -53,6 +51,53 @@ class IRMetadata:
             feature_columns=data["feature_columns"],
             units=data.get("units", {}),
             source_url=data.get("source_url", ""),
-            transform_history=data.get("transform_history", []),
             extra=data.get("extra", {}),
         )
+
+
+@dataclass
+class WindowConfig:
+    """
+    Configuration for sliding window creation.
+
+    Attributes:
+        input_length: Number of past timesteps (L) to use as input.
+        horizon: Number of future timesteps (H) to predict.
+        y_start: Gap between input end and target start. Default is 1,
+            meaning target starts immediately after input.
+        input_columns: Feature columns to use for input. None means all columns.
+        target_columns: Feature columns to predict. None means same as input_columns.
+    """
+
+    input_length: int = 12
+    horizon: int = 12
+    y_start: int = 1
+    input_columns: list[str] | None = None
+    target_columns: list[str] | None = None
+
+
+@dataclass
+class SplitConfig:
+    """
+    Configuration for temporal train/val/test split.
+
+    The test ratio is computed as 1 - train_ratio - val_ratio.
+
+    Attributes:
+        train_ratio: Fraction of data for training (e.g., 0.7 for 70%).
+        val_ratio: Fraction of data for validation (e.g., 0.1 for 10%).
+    """
+
+    train_ratio: float = 0.7
+    val_ratio: float = 0.1
+
+    def __post_init__(self):
+        if self.train_ratio + self.val_ratio >= 1.0:
+            raise ValueError("train_ratio + val_ratio must be less than 1.0")
+        if self.train_ratio <= 0 or self.val_ratio < 0:
+            raise ValueError("Ratios must be positive")
+
+    @property
+    def test_ratio(self) -> float:
+        """Compute the test ratio as the remainder."""
+        return 1.0 - self.train_ratio - self.val_ratio
