+++ datasets/beijing_air.py
@@ -15,7 +15,7 @@ import pandas as pd
 
 from gnn_benchmark.core.types import DatasetInfo
 from gnn_benchmark.datasets.base import DatasetLoader
-from gnn_benchmark.utils.graph import haversine_distance
+from gnn_benchmark.utils.geo import haversine_distance
 
 # Node orders for different subdivisions
 BEIJING_NODE_ORDER = [
